fix: recursive buildCustomElementConstructor
