+++ packages/@lwc/engine/dom/src/apis/build-custom-element-constructor.ts
@@ -1,0 +1,40 @@
+/*
+ * Copyright (c) 2018, salesforce.com, inc.
+ * All rights reserved.
+ * SPDX-License-Identifier: MIT
+ * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/MIT
+ */
+
+import { LightningElement } from './lightning-element';
+
+type ComponentConstructor = typeof LightningElement;
+type HTMLElementConstructor = typeof HTMLElement;
+
+/**
+ * This function builds a Web Component class from a LWC constructor so it can be
+ * registered as a new element via customElements.define() at any given time.
+ *
+ * @deprecated since version 1.3.11
+ *
+ * @example
+ * ```
+ * import { buildCustomElementConstructor } from 'lwc';
+ * import Foo from 'ns/foo';
+ * const WC = buildCustomElementConstructor(Foo);
+ * customElements.define('x-foo', WC);
+ * const elm = document.createElement('x-foo');
+ * ```
+ */
+export function deprecatedBuildCustomElementConstructor(
+    Ctor: ComponentConstructor
+): HTMLElementConstructor {
+    if (process.env.NODE_ENV !== 'production') {
+        /* eslint-disable-next-line no-console */
+        console.warn(
+            'Deprecated function called: "buildCustomElementConstructor" function is deprecated and it will be removed.' +
+                `Use "${Ctor.name}.CustomElementConstructor" static property of the component constructor to access the corresponding custom element constructor instead.`
+        );
+    }
+
+    return Ctor.CustomElementConstructor;
+}
