+++ packages/@lwc/engine/dom/src/index.ts
@@ -17,7 +17,6 @@ export {
     wire,
     readonly,
     unwrap,
-    buildCustomElementConstructor,
     setFeatureFlag,
     setFeatureFlagForTest,
     registerTemplate,
@@ -29,6 +28,7 @@ export {
 } from '../../src';
 
 // Public APIs -------------------------------------------------------------------------------------
+export { deprecatedBuildCustomElementConstructor as buildCustomElementConstructor } from './apis/build-custom-element-constructor';
 export { createElement } from './apis/create-element';
 export { getComponentConstructor } from './apis/get-component-constructor';
 export { isNodeFromTemplate } from './apis/is-node-from-template';
